requirements.txt
fastapi
fastapi_limiter
python-decouple
redis
